requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-decouple
black